finished up testing for now

+++ mediators/synchronizer.py
@@ -1,3 +1,5 @@
+import copy
+
 from models.problem import Problem
 from models.server import Server
 
@@ -23,44 +25,67 @@ class Synchronizer:
 
     # this ensures that the models store the same data as the buckets
     # adds a problem to where it belongs. ie the server and proper bucket
-    def addProblem(self, problem: Problem):
+    def addProblem(self, problem: Problem) -> bool:
         pid = problem.problemID
         sid = problem.serverID
-
+        
+        if sid not in self.servers:
+            # print("Server ID not found in servers.")
+            return False
+        
         server = self.servers[sid]
         
+        backupProblems = copy.deepcopy(server.problems)
+        backupBucket = copy.deepcopy(self.problemBucket.buckets)
+        
         # we have a problem with this id already, thus we need to remove it
-        if server.problems[pid] is not None:
+        if problem in server.problems:
             problemToRemove = server.problems[pid]
-            if not self.dowBucket.removeFromBucket(problemToRemove):
+            if not self.problemBucket.removeFromBucket(problemToRemove):
+                # print("Failed to remove problem from bucket.")
                 return False # failed to remove
         
         # we either didn't have a problem to remove, or did and we removed it
         # we can safely add to the server and the buckets now
         if not server.addProblem(problem):
+            # print("Failed to add problem to server.")
+            # if we fail to add to the server, we should revert the bucket
+            # since we just removed a problem from it
+            self.problemBucket.buckets = backupBucket
             return False # failed to add to server
-        
-        if not self.dowBucket.addToBucket(problem):
-            server.removeProblem(problem) # since we can't add to server
+
+        if not self.problemBucket.addToBucket(problem):
+            # since we failed to add a problem to the bucket, we must revert the server
+            server.problems = backupProblems
+            # print("Failed to add problem to bucket.")
             return False # failed to add to bucket
         
         return True
 
     # removes a problem from the server and the bucket
-    def removeProblem(self, problem: Problem):
-        pid = problem.problemID
+    def removeProblem(self, problem: Problem) -> bool:
         sid = problem.serverID
+        
+        if sid not in self.servers:
+            # print("Server ID not found in servers.")
+            return False
+        
         server = self.servers[sid]
         
+        backupProblems = copy.deepcopy(server.problems)
+        backupBucket = copy.deepcopy(self.problemBucket.buckets)
+        
         # we have a problem with this id already, thus we need to remove it
-        if server.problems[pid] is None:
+        if problem not in server.problems:
             return False # nothing to remove
 
         if not server.removeProblem(problem):
+            # we failed to remove the problem, but we dont have to revert
             return False
         
-        if not self.dowBucket.removeFromBucket(problem):
-            server.addProblem(problem) # add it back to the server, to keep it aligned
+        if not self.problemBucket.removeFromBucket(problem):
+            server.problems = backupProblems
+            self.problemBucket.buckets = backupBucket
             return False
         
         return True
@@ -72,32 +97,48 @@ class Synchronizer:
     def changeAlertIntervals(self, serverID: int, newIntervals: list[int]) -> bool:
         if serverID not in self.servers:
             return False
-        
+                
         server = self.servers[serverID]
         
+        # incase we encounter an error, we can revert
+        backupIntervals = copy.deepcopy(server.settings.contestTimeIntervals)
+        backupBucket = copy.deepcopy(self.contestTimeBucket.buckets)
+        
         # get and remove the current intervals
-        currentIntervals = server.settings.contestTimeAlerts
-        for interval in currentIntervals:
-            if not self.contestTimeBucket.removeFromBucket(interval, serverID):
-                return False
+        currentIntervals = server.settings.contestTimeIntervals
+        if currentIntervals and len(currentIntervals) != 0:
+            for interval in currentIntervals:
+                if not self.contestTimeBucket.removeFromBucket(interval, serverID):
+                    server.settings.contestTimeIntervals = backupIntervals
+                    self.contestTimeBucket.buckets = backupBucket
+                    # print("Failed to remove contest time interval from bucket.")
+                    return False
         
         # add the new intervals
         for interval in newIntervals:
             if not self.contestTimeBucket.addToBucket(interval, serverID):
+                server.settings.contestTimeIntervals = backupIntervals
+                self.contestTimeBucket.buckets = backupBucket
+                # print("Failed to add contest time interval to bucket.")
                 return False
+        
+        server.settings.contestTimeIntervals = newIntervals
+        server.toJSON() # save after we update the setting
         
         return True
     
     def changeContestAlertParticpation(self, serverID: int, participate: bool) -> bool:
         if serverID not in self.servers:
+            # print("Server ID not found in servers.")
             return False
 
         server = self.servers[serverID]
         serverSettings = server.settings
-
-        serverSettings.contestTimeAlerts = participate
         
-        currentIntervals = server.settings.contestTimeAlerts
+        backupSettings = copy.deepcopy(serverSettings)
+        backupBucket = copy.deepcopy(self.contestTimeBucket.buckets)
+
+        currentIntervals = server.settings.contestTimeIntervals
         if not currentIntervals:
             currentIntervals = [15] # set to a default of 15 mins is empty
             
@@ -105,40 +146,74 @@ class Synchronizer:
         # otherwise, remove them from the bucket 
         for interval in currentIntervals:
             if participate:
-                self.contestTimeBucket.addToBucket(interval, serverID)
+                if not self.contestTimeBucket.addToBucket(interval, serverID):
+                    server.settings = backupSettings
+                    self.contestTimeBucket.buckets = backupBucket
+                    # print("Failed to add contest time interval to bucket.")
+                    return False
             else:
-                self.contestTimeBucket.removeFromBucket(interval, serverID)
-    
+                if not self.contestTimeBucket.removeFromBucket(interval, serverID):
+                    server.settings = backupSettings
+                    self.contestTimeBucket.buckets = backupBucket
+                    # print("Failed to remove contest time interval from bucket.")
+                    return False
+                
+        serverSettings.contestTimeAlerts = participate
+        server.toJSON() # save after we update the setting
+
+        return True
+
     # ==============================================
     # STATIC ALERTS
     # ==============================================
     
     def changeStaticAlert(self, serverID: int, alert: StaticTimeAlert, participate: bool) -> bool:
         if serverID not in self.servers:
+            # print("Server ID not found in servers.")
             return False
         
         # update the server settings
         server = self.servers[serverID]
         serverSettings = server.settings
         
+        backupSettings = copy.deepcopy(serverSettings)
+        backupBucket = copy.deepcopy(self.staticTimeBucket.buckets)
+        
+        # update the setting if and only if it is not already set to 
+        # what it is trying to be changed to 
         match (alert):
             case (StaticTimeAlert.WEEKLY_CONTEST):
+                if serverSettings.weeklyContestAlerts == participate:
+                    return True
                 serverSettings.weeklyContestAlerts = participate
             case (StaticTimeAlert.BIWEEKLY_CONTEST):
+                if serverSettings.biweeklyContestAlerts == participate:
+                    return True
                 serverSettings.biweeklyContestAlerts = participate
             case (StaticTimeAlert.DAILY_PROBLEM):
+                if serverSettings.officialDailyAlerts == participate:
+                    return True
                 serverSettings.officialDailyAlerts = participate
             case _:
+                print("Invalid alert")
                 return False
-            
-        server.toJSON() # save after we update the setting
         
+        # if we got here, then we must update a bucket to reflect the change        
         # change the bucket to reflect
         # on participate, we want to add
         # if we do NOT want to participate, then we must remove
         if participate:
-            return self.staticTimeBucket.addToBucket(alert, serverID)
+            if not self.staticTimeBucket.addToBucket(alert, serverID):
+                server.settings = backupSettings
+                self.staticTimeBucket.buckets = backupBucket
+                print("Failed to add static time alert to bucket.")
+                return False
         else:
-            return self.staticTimeBucket.removeFromBucket(alert, serverID)
-            
+            if not self.staticTimeBucket.removeFromBucket(alert, serverID):
+                server.settings = backupSettings
+                self.staticTimeBucket.buckets = backupBucket
+                print("Failed to remove static time alert from bucket.")
+                return False
         
+        server.toJSON() # save after we update the setting
+        return True
